refactor: extract reference counting to Referenced base class
